PROJ-25_datePicker refactor: split code into files and add index.ts files for isolation
