+++ src/components/DatePicker/DatePicker.module.scss
@@ -59,91 +59,14 @@
 
   // if date picker has an error
   &.error {
-    span {
+    & > span {
       color: var(--color-danger-500);
     }
 
     .dateDisplayer {
-      cursor: default;
       color: var(--color-danger-500);
       background-color: var(--color-dark-500);
       border-color: var(--color-danger-500);
-    }
-  }
-
-  .calendar {
-    position: absolute;
-    z-index: 10;
-    top: 63px;
-
-    display: flex;
-    flex-direction: column;
-    justify-content: space-between;
-
-    width: 300px;
-    height: 348px;
-    padding: 16px 24px;
-
-    background-color: var(--color-dark-500);
-    border: 1px solid var(--color-dark-300);
-    border-radius: 2px;
-    outline: none;
-
-    &.calendarWithoutLabel {
-      top: 39px;
-    }
-
-    .monthsSettings {
-      display: flex;
-      align-items: center;
-      justify-content: space-between;
-
-      span {
-        @include Bold_text_16;
-
-        margin: 6px 0 6px 2px;
-        color: var(--color-light-100);
-      }
-
-      button {
-        all: unset;
-
-        cursor: pointer;
-
-        display: inline-flex;
-        align-items: center;
-        justify-content: center;
-
-        width: 36px;
-        height: 36px;
-
-        background-color: var(--color-dark-100);
-        border-radius: 50%;
-      }
-    }
-
-    .daysInWeek {
-      display: flex;
-      justify-content: space-between;
-
-      div {
-        @include Regular_text_16;
-
-        display: flex;
-        align-items: center;
-        justify-content: center;
-
-        width: 36px;
-        height: 40px;
-
-        color: var(--color-light-900);
-      }
-    }
-
-    .calendarDays {
-      display: grid;
-      grid-template-columns: repeat(7, 1fr);
-      grid-template-rows: repeat(6, 1fr);
     }
   }
